+++ starts/worker.go
@@ -80,11 +80,7 @@ func CreateFlowBatches(ctx context.Context, db *sqlx.DB, rp *redis.Pool, ec *ela
 
 	// if we are meant to create a new contact, do so
 	if start.CreateContact() {
-		if org == nil {
-			org, err = models.GetOrgAssets(ctx, db, start.OrgID())
-			if err != nil {
-				return errors.Wrapf(err, "error loading org assets")
-			}
+		if assets == nil {
 			assets, err = models.GetSessionAssets(org)
 			if err != nil {
 				return errors.Wrapf(err, "error loading session assets")